UI new palette for MuiTheme
