+++ frontend/src/graphs/styledGraphComponents/LinePlot.js
@@ -7,7 +7,7 @@ import {
   mainTextColor,
   dashboardChartAxisColor,
   dashboardChartGridColor,
-} from '../../utils/oscarMuiTheme';
+} from '../../muiTheme/colors';
 
 import ProgressIndicator from '../ProgressIndicator';
 import EmptyChartMessage from '../EmptyChartMessage';
