+++ frontend/src/muiTheme/colors.js
@@ -1,0 +1,37 @@
+export const mainBoxShadow = '0px 1px 34px 2px rgba(39, 44, 55, 0.41)';
+
+// needs to be sorted out
+
+export const activeButtonBackgroundColor = 'rgb(21, 25, 34)';
+export const muiSelectBackgroundColor = 'rgb(70,75,88)';
+
+export const dashboardChartAxisColor = 'rgba(255, 255, 255, 0.64)';
+export const dashboardCardGreyColor = 'rgba(255, 255, 255, 0.3)';
+export const dashboardChartGridColor = 'rgba(248, 248, 248, 0.11)';
+export const dashboardChartRedColor = 'rgb(243, 62, 62)';
+export const dashboardChartRedColorTransparent = 'rgba(243, 62, 62, 0.2)';
+export const dashboardChartYellowColor = 'rgb(220, 195, 64)';
+export const dashboardChartBlueColor = 'rgb(64, 70, 220)';
+export const dashboardChartPurpleColor = 'rgb(229, 62, 243)';
+export const dashboardChartLightBlueColor = 'rgb(64, 220, 220)';
+export const iconColor = 'rgb(248, 248, 248)';
+
+export const mainTextColor = 'rgb(248, 248, 248)';
+export const lightTextColor = 'rgb(242, 242, 242)';
+export const greyTextColor = 'rgba(248, 248, 248, 0.7)';
+export const lightGreyTextColor = 'rgba(248, 248, 248, 0.39)';
+export const accentColor = {
+  main: 'rgb(243, 150, 62)',
+};
+
+export const mainBackgroundColor = 'rgb(56, 63, 81)';
+export const backgroundColor = 'rgb(53, 60, 77)';
+export const buttonBackgroundColor = 'rgb(34, 40, 54)';
+export const dropdownBackgroundColor = 'rgb(34, 40, 54)';
+export const inputBackgroundColor = 'rgba(0,0,0,0.2)';
+export const autocompleteBackgroundColor = 'rgb(56, 63, 81)';
+export const muiListItemBackgroundColor = 'rgba(0, 0, 0, 0.14)';
+
+export const inputColor = 'rgba(248, 248, 248, 0.89)';
+export const dashboardDateColor = 'rgba(248, 248, 248, 0.38)';
+export const dashboardChartGreenColor = 'rgb(87, 243, 62)';
