+++ frontend/src/screens/Map/MapButtons/ActionButtons.js
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 
-import { mainBackgroundColor } from '../../../utils/oscarMuiTheme';
+import { mainBackgroundColor } from '../../../muiTheme/colors';
 import logo from '../../../data/logo_orange.png';
 import GeolocationButton from './GeolocationButton';
 import LayerButton from './LayersButton';
