+++ frontend/src/screens/Map/MapButtons/LayersButton.js
@@ -8,7 +8,7 @@ import Fab from '@material-ui/core/Fab';
 
 import { actions as mapActions } from '../../../actions';
 import { SELECT_LAYERS } from '../../../constants/currentActivity';
-import { lightTextColor, accentColor } from '../../../utils/oscarMuiTheme';
+import { lightTextColor, accentColor } from '../../../muiTheme/colors';
 
 export const LayerButton = ({ isMobile, selectingLayers, cancelSelecting, selectLayers }) => {
   const mobileStyle = {
